pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "structui"
version = "0.1.0"
description = "A format-agnostic, schema-driven, hierarchical configuration UI."
readme = "README.md"
authors = [
    { name = "structui contributors" }
]
license = { text = "MIT" }
requires-python = ">=3.9"
dependencies = [
    "nicegui>=1.4.0",
    "pyyaml>=6.0.1",
    "pywin32>=300; sys_platform == 'win32'"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12"
]

[project.scripts]
structui = "structui.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
